fix(compliance): use published SDK storyboard routing
