+++ .agents/sdk-shim-ledger.json
@@ -30,7 +30,7 @@
     "paths": [
       "scripts/overlay-compliance-cache.sh",
       "scripts/run-storyboards-matrix.sh",
-      "tests/patch-sdk-rc15.test.cjs"
+      "tests/sdk-runtime-compat.test.cjs"
     ],
     "terms": [
       "node_modules/@adcp/sdk/compliance/cache",
@@ -38,27 +38,6 @@
       "node_modules/@adcp/sdk/dist/lib/types/schemas.generated.js",
       "node_modules/@adcp/sdk/dist/lib/types/schemas.generated.mjs",
       "schemas.generated.js"
-    ]
-  },
-  {
-    "id": "sdk-storyboard-product-wire",
-    "title": "SDK storyboard product-wire routing fix",
-    "kind": "installed_sdk_runtime_patch",
-    "status": "temporary",
-    "owner": "training-agent",
-    "upstream": "adcontextprotocol/adcp-client#2495, adcontextprotocol/adcp-client#2527",
-    "problem": "The rc.18 storyboard task map forces get_products onto a legacy-only wire, and scoped transports discover capabilities without carrying them into request adaptation. The latter makes 3.1 sellers look pre-3.1 and strips filters.pricing_currencies.",
-    "localBehavior": "Patch both compiled SDK module formats after installation so the storyboard runner leaves get_products wire selection unspecified and request-local capabilities drive version adaptation without cross-tenant caching.",
-    "removalCondition": "Remove the patch script, Docker invocations, and ledger entry after an SDK release fixes storyboard get_products routing and carries scoped capabilities into request adaptation, and both current and released-3.0 storyboard matrices pass without the patch.",
-    "paths": [
-      "scripts/patch-sdk-rc15.mjs",
-      "tests/patch-sdk-rc15.test.cjs"
-    ],
-    "terms": [
-      "node_modules/@adcp/sdk/dist/lib/testing/storyboard/task-map.js",
-      "node_modules/@adcp/sdk/dist/lib/testing/storyboard/task-map.mjs",
-      "node_modules/@adcp/sdk/dist/lib/core/SingleAgentClient.js",
-      "node_modules/@adcp/sdk/dist/lib/core/SingleAgentClient.mjs"
     ]
   }
 ]
