+++ static/compliance/source/protocols/media-buy/scenarios/canonical_formats.yaml
@@ -340,6 +340,7 @@ phases:
         response_schema_ref: "media-buy/get-products-response.json"
         doc_ref: "/media-buy/task-reference/get_products"
         comply_scenario: schema_compliance
+        response_projection: raw
         stateful: true
         expected: |
           Return products including the seeded canonical_formats_mrec_display
@@ -445,6 +446,7 @@ phases:
         response_schema_ref: "media-buy/get-products-response.json"
         doc_ref: "/media-buy/task-reference/get_products"
         comply_scenario: schema_compliance
+        response_projection: raw
         stateful: true
         expected: |
           Return the seeded canonical_formats_ids_only_display product with a
@@ -495,6 +497,7 @@ phases:
         response_schema_ref: "media-buy/get-products-response.json"
         doc_ref: "/media-buy/task-reference/get_products"
         comply_scenario: schema_compliance
+        response_projection: raw
         stateful: true
         expected: |
           Return the seeded canonical_formats_options_only_display product with
@@ -553,6 +556,7 @@ phases:
         response_schema_ref: "media-buy/get-products-response.json"
         doc_ref: "/media-buy/task-reference/get_products"
         comply_scenario: schema_compliance
+        response_projection: raw
         stateful: true
         expected: |
           Return the seeded canonical_formats_custom_takeover product with a
@@ -619,6 +623,7 @@ phases:
         response_schema_ref: "media-buy/get-products-response.json"
         doc_ref: "/media-buy/task-reference/get_products"
         comply_scenario: schema_compliance
+        response_projection: raw
         stateful: true
         expected: |
           Return the seeded canonical_formats_experimental_display product with
@@ -683,6 +688,7 @@ phases:
         response_schema_ref: "media-buy/get-products-response.json"
         doc_ref: "/media-buy/task-reference/get_products"
         comply_scenario: schema_compliance
+        response_projection: raw
         auth:
           type: api_key
           from_test_kit: true
@@ -744,6 +750,7 @@ phases:
         response_schema_ref: "media-buy/get-products-response.json"
         doc_ref: "/media-buy/task-reference/get_products"
         comply_scenario: schema_compliance
+        response_projection: raw
         stateful: true
         expected: |
           Return the seeded canonical_formats_mrec_display product so create-time
